# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open API - E Commerce",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "api/payment/rating": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create Rating for product",
                "description": "Create Rating for product",
                "operationId": "c0f9ee1a8ebfb8b76f139d93142bf504",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "rate": {
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "order_id": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "rate": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "customer_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
...
}